requirements.txt
# Core ML and NLP packages
torch
transformers

# Data analysis packages
pandas
numpy

# API and web packages
requests
yfinance
python-dotenv

# Email functionality
email-validator
